fix: generate InfraRunOptions from terraform show
- We'd run into problems if we are generating the infra options from a
  DeploymentInventory if the environment had issues while spining up
- Thus get all the values for the InfraRunOptions from the
  terraform::show command
